Added a compressed DB field.
The compressed DB field will automatically deflate the data before
storage in the DB and inflate again when retrieving the data. This is
done transparently for the end user. This is good if you need to store
large body of text and do not need to search into it with the SQL
commands.